vista parcial para mostrar el rating de los cursos y atributos pesonalizados eloquent

+++ database/migrations/2019_03_31_051504_create_reviews_table.php
@@ -20,7 +20,7 @@ class CreateReviewsTable extends Migration
             $table->unsignedInteger('user_id');
             $table->foreign('user_id')->references('id')->on('courses');
             //presicion de 2
-            $table->float('rating',2);
+            $table->integer('rating');
             $table->text('comment')->nullable();
 
             $table->timestamps();
